+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Bug-fixes
 
+- [x] [#560 - Error on "npm run android"](https://github.com/scriptPilot/app-framework/issues/560)
 - [x] [#575 - Icons not updated on image background color change](https://github.com/scriptPilot/app-framework/issues/575)
 
 ## Version 1.7.23